# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DeletionResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "description": "How many analyses were removed.",
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeletionResponse",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "Fatal conditions, one per HTTP status the API can return.\n\n| code                    | status |\n|-------------------------|--------|\n| file_too_large          | 413    |\n| unsupported_media_type  | 415    |\n| animated_image          | 415    |\n| corrupt_image           | 422    |\n| image_too_small         | 422    |\n| out_of_distribution     | 422    |\n| unauthenticated         | 401    |\n| not_found               | 404    |\n| rate_limited            | 429    |\n| not_implemented         | 501    |\n| service_degraded        | 503    |",
        "enum": [
          "file_too_large",
          "unsupported_media_type",
          "animated_image",
          "corrupt_image",
          "image_too_small",
          "out_of_distribution",
          "unauthenticated",
          "not_found",
          "rate_limited",
          "not_implemented",
          "service_degraded"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "description": "Human-readable explanation. Safe to display to a user.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Correlates this failure with the server logs.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Every non-2xx response from this API has exactly this body.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryItem": {
        "additionalProperties": true,
        "description": "One past analysis, as stored.\n\nA row, not a re-serialised `AnalyzeResponse`: the two can legitimately differ\nonce the schema evolves, and pretending an old row is a current response\nwould quietly rewrite history.",
        "properties": {
          "calibrated": {
            "default": false,
            "title": "Calibrated",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "domain_confidence": {
            "title": "Domain Confidence",
            "type": "number"
          },
          "findings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "specialist_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialist Model"
          },
          "vlm_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Description"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "required": [
          "id",
          "created_at",
          "domain",
          "domain_confidence"
        ],
        "title": "HistoryItem",
        "type": "object"
      },
...
      "RequestHistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "title": "Items",
            "type": "array"
          },
          "retention_days": {
            "description": "Stored images and rows are deleted after this many days.",
            "title": "Retention Days",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count",
          "retention_days"
        ],
        "title": "RequestHistoryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDamage analysis that reports what it cannot do.\n\nAn uploaded photograph passes a **gate** (is this a damage photo at all?), then a\n**router** (which domain?), and then either a **domain specialist** -- if one has\nbeen trained -- or an honest fallback that returns no findings and says why.\n\nConfidence values are either calibrated and marked `calibrated: true`, or not\ncalibrated and marked `false`. There is no third state.\n",
    "title": "BioVision",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/requests": {
      "delete": {
        "description": "Erase the caller's history and every image stored with it.\n\nThe \"delete all my data\" path. It exists in v1 because the README's retention\nclaim is only meaningful alongside a way to act on it sooner.",
        "operationId": "delete_all_requests_v1_requests_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign-in required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete every analysis belonging to the caller",
        "tags": [
          "history"
        ]
      },
      "get": {
        "description": "Return the caller's own past analyses, newest first.",
        "operationId": "list_requests_v1_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign-in required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The authenticated user's own request history",
        "tags": [
          "history"
        ]
      }
    },
    "/v1/requests/{analysis_id}": {
      "delete": {
        "description": "Delete one analysis and its stored image.\n\nAnother user's analysis is invisible under RLS, so a request naming one\ndeletes nothing and returns 404 -- identical to a genuinely absent id. That\nis deliberate: distinguishing the two would confirm the id exists.",
        "operationId": "delete_request_v1_requests__analysis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sign-in required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such analysis for this caller"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete one of the caller's analyses",
        "tags": [
          "history"
        ]
      }
    }
  }
}